cleanup package structure
