+++ lib/oscillators.dart
@@ -1,0 +1,3 @@
+abstract class SineOscillator {
+  double next();
+}
