+++ lib/wasm_osc/sine_osc.dart
@@ -1,0 +1,15 @@
+import '../oscillators.dart';
+import 'wasm_helper.dart';
+
+class WasmSineOscillator implements SineOscillator {
+  late final WasmOscillator _wasmOsc;
+
+  WasmSineOscillator(int sampleRate, double frequency) {
+    _wasmOsc =
+        WasmOscillator('../../as-audio/build/osc.wasm', 'SineOscillator');
+    _wasmOsc.frequency = frequency;
+  }
+
+  @override
+  double next() => _wasmOsc.nextSample();
+}
